Rewrite the readiness queue
The current readiness queue implication has some concurrency bugs. So,
instead of fixing the bug, let's rewrite everything! :)

But more seriously, the current implication has some problems (besides
having bugs):

* It's complicated
* It is not fair when using level-triggered notifications
* It is !Sync

This implication will hopefully resolve the bugs, as well as simplifying
the implementation by reducing all the critical concurrent state for a
given node to a single atomic variable. It also changes the queueing
strategy from a stack to an actual queue based on the 1024cores MPSC
queue. However, this queue is able to reuse the nodes.