Coordinate calls to `Poll::poll`
The last step in making `Poll` fully `Sync` is to coordinate calls to
`poll`. There can only be one single concurrent call to `poll`.

In order to respect the timeout argument to `poll`, Mutex cannot be used
for coordination. This is due to the lack of a `lock_timeout` function.
Instead, we use a combination of an atomic, a mutex, and a condvar. The
atomic enables a fast path when `Poll` is used on a single thread. In
the event of concurrent calls to `poll`, extra threads block using the
condvar until they are granted access.